refactor: extract domain models with intent-revealing names
Split resolve.js (507 lines) into focused model files:
- model/api-resource.js: settle → mergeResourceProperties, derive → buildResourceDefinitions
- model/event-resource.js: createEventResource
- model/entity-type.js: createEntityTypes
- model/group.js: createGroup, formatServiceTitle
- model/visibility.js: determineVisibility (not "resolve")
- model/naming.js: stripNamespacePrefix, buildGroupId, findPackageId
- model/package.js: createProducts, createPackages

Naming conventions:
- create*: produces new domain objects
- extract*: parses information from strings
- determine*: makes decisions based on rules
- build*: constructs ID/URL strings
- find*: searches through collections
- collect*: traverses and gathers

resolve.js is now 67 lines of pure orchestration.
