+++ lib/compiler/emit.js
@@ -2,8 +2,8 @@ const _ = require("lodash");
 const defaults = require("../defaults");
 const { compareAndHandleCustomORDContentWithExistingContent } = require("../extend-ord-with-custom");
 
-function emit(ir) {
-    const { config, resolved, customOrd, extensions } = ir;
+function emit(ordDocument) {
+    const { config, resolved, customOrd, extensions } = ordDocument;
 
     let document = {
         $schema: defaults.$schema,
