+++ lib/compiler/index.js
@@ -4,8 +4,8 @@ const { validate } = require("./validate");
 const { emit } = require("./emit");
 
 function compile(csn, extensions = []) {
-    const ir = parse(csn, extensions);
-    const resolved = resolve(ir);
+    const document = parse(csn, extensions);
+    const resolved = resolve(document);
     validate(resolved);
     return emit(resolved);
 }
