+++ lib/compiler/validate.js
@@ -1,14 +1,12 @@
-const Logger = require("../logger");
-
-function validate(ir) {
+function validate(document) {
     const errors = [];
     const seenOrdIds = new Set();
 
     const allResources = [
-        ...(ir.resolved.apiResources || []),
-        ...(ir.resolved.eventResources || []),
-        ...(ir.resolved.entityTypes || []),
-        ...(ir.resolved.integrationDependencies || []),
+        ...(document.resolved.apiResources || []),
+        ...(document.resolved.eventResources || []),
+        ...(document.resolved.entityTypes || []),
+        ...(document.resolved.integrationDependencies || []),
     ];
 
     for (const resource of allResources) {
